# swagger.yaml
openapi: 3.0.4
info:
  title: LAPEN Agenda API
  description: Tennis Court Management System with Betting Integration
  version: 1.0.0
  contact:
    email: lapen.ptc@gmail.com
servers:
  - url: http://localhost:5001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminSession:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Tournament:
      type: object
      properties:
        id: {type: integer, example: 1}
        name: {type: string, example: "Copa LAPEN 2026"}
        slug: {type: string, example: "copa-lapen-2026"}
        description: {type: string, nullable: true}
        location: {type: string, nullable: true}
        start_date: {type: string, format: date, example: "2026-11-10"}
        end_date: {type: string, format: date, example: "2026-11-12"}
        registration_opens_at: {type: string, format: date-time, nullable: true, description: "Brazil local time, no timezone"}
        registration_closes_at: {type: string, format: date-time, nullable: true, description: "Brazil local time, no timezone"}
        rules_text: {type: string, nullable: true}
        contact_info: {type: string, nullable: true}
        match_format:
          type: string
          enum: [best_of_3_super_tb, pro_set_8, single_set_6]
          description: Chosen when the tournament is created, locked after the first result
        no_ad: {type: boolean}
        match_tiebreak_points: {type: integer, example: 10}
        status:
          type: string
          enum: [draft, registration_open, registration_closed, in_progress, finished, cancelled]
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
paths:
...
  /api/admin/tournaments:
    get:
      tags: [Tournaments]
      summary: List tournaments (active first) with category and pending-registration counts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tournaments
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournaments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Tournament'
                        - type: object
                          properties:
                            categories_count: {type: integer}
                            pending_registrations: {type: integer}
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    post:
      tags: [Tournaments]
      summary: Create a tournament (always starts as draft)
      description: The slug is generated from the name unless given. Defaults match_format best_of_3_super_tb, no_ad true, match_tiebreak_points 10.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, start_date, end_date]
              properties:
                name: {type: string}
                slug: {type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
                description: {type: string}
                location: {type: string}
                start_date: {type: string, format: date}
                end_date: {type: string, format: date}
                registration_opens_at: {type: string, example: "2026-11-01T08:00"}
                registration_closes_at: {type: string, example: "2026-11-08T18:00"}
                rules_text: {type: string}
                contact_info: {type: string}
                match_format: {type: string, enum: [best_of_3_super_tb, pro_set_8, single_set_6]}
                no_ad: {type: boolean}
                match_tiebreak_points: {type: integer}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Slug already taken (code slug_taken)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: Tournament with its categories and registration counts
      description: Registrations whose deadline passed are closed lazily when this is read.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
                  categories:
                    type: array
                    items: {$ref: '#/components/schemas/TournamentCategory'}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    put:
      tags: [Tournaments]
      summary: Update a tournament (partial)
      description: The slug never changes. Finished or cancelled tournaments are read-only. match_format is locked after the first result (code format_locked).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of the fields accepted on creation, except slug
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Read-only or format locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    delete:
      tags: [Tournaments]
      summary: Delete a draft tournament (cascades to categories)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Not a draft (code not_draft)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/status:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    put:
      tags: [Tournaments]
      summary: Change the tournament status
      description: |
        Allowed moves: draft to registration_open or cancelled, registration_open to registration_closed or cancelled,
        registration_closed to registration_open, in_progress or cancelled, in_progress to finished or cancelled.
        Only one tournament can be active (registration_open, registration_closed, in_progress) at a time:
        opening a second one returns 409 with code active_tournament_exists and the blocking tournament in active_tournament.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: {type: string, enum: [registration_open, registration_closed, in_progress, finished, cancelled]}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
        '400':
          description: Unknown status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Invalid transition or guard failed (codes invalid_transition, no_categories, active_tournament_exists, registration_window_past, draw_published, no_draw, categories_unfinished)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/categories:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    post:
      tags: [Tournaments]
      summary: Create a category
      description: groups_knockout needs group_target_size (3 or 4), qualifiers_per_group (1 or 2) and min_entries of at least 6. Group settings are dropped for other formats.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string}
                draw_format: {type: string, enum: [knockout, round_robin, groups_knockout], default: knockout}
                min_entries: {type: integer, default: 4}
                max_entries: {type: integer, nullable: true}
                group_target_size: {type: integer, enum: [3, 4]}
                qualifiers_per_group: {type: integer, enum: [1, 2]}
                num_seeds: {type: integer, default: 0}
                wo_tolerance_min: {type: integer, default: 15}
                min_rest_min: {type: integer, default: 30}
                sort_order: {type: integer}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  category: {$ref: '#/components/schemas/TournamentCategory'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Duplicate name (category_name_taken) or tournament locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/categories/{category_id}:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
      - {name: category_id, in: path, required: true, schema: {type: integer}}
    put:
      tags: [Tournaments]
      summary: Update a category (partial)
      description: Format, group settings and num_seeds are locked once the draw exists (code draw_exists). max_entries cannot drop below the confirmed count and num_seeds cannot drop below an assigned seed.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of the fields accepted on creation
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  category: {$ref: '#/components/schemas/TournamentCategory'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Locked, below_confirmed, seed_assigned or duplicate name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    delete:
      tags: [Tournaments]
      summary: Delete a category without registrations and without a draw
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Has registrations or draw exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/registrations:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: List registrations with per-category counts
      description: Includes e-mail and phone (admin only). Search is accent-insensitive on name, display name, e-mail and phone.
      security:
        - BearerAuth: []
      parameters:
        - {name: category_id, in: query, schema: {type: integer}}
        - {name: status, in: query, schema: {type: string, enum: [pending, confirmed, rejected, cancelled, waitlist, withdrawn]}}
        - {name: q, in: query, schema: {type: string}}
      responses:
        '200':
          description: Registrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  registrations:
                    type: array
                    items: {$ref: '#/components/schemas/TournamentRegistration'}
                  counts:
                    type: object
                    description: category id to registrations per status
                    additionalProperties:
                      type: object
                      additionalProperties: {type: integer}
        '400':
          description: Invalid status filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    post:
      tags: [Tournaments]
      summary: Register someone manually (for example received by WhatsApp)
      description: Starts as confirmed unless status is pending. Respects the category capacity. Optional user_id links an approved LAPEN member.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, full_name, email, phone]
              properties:
                category_id: {type: integer}
                full_name: {type: string}
                display_name: {type: string}
                email: {type: string}
                phone: {type: string}
                notes: {type: string}
                status: {type: string, enum: [pending, confirmed], default: confirmed}
                user_id: {type: integer}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration: {$ref: '#/components/schemas/TournamentRegistration'}
        '400':
          description: Invalid data or user is not an approved LAPEN member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Category full, duplicate e-mail or member, draw exists, tournament locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/registrations/batch:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    post:
      tags: [Tournaments]
      summary: Confirm, reject or cancel several registrations
      description: Always answers 200 with the items that worked and the ones that failed (each with error and code).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, status]
              properties:
                ids: {type: array, maxItems: 200, items: {type: integer}}
                status: {type: string, enum: [confirmed, rejected, cancelled]}
                rejection_reason: {type: string}
      responses:
        '200':
          description: Per-item result
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: array
                    items: {type: integer}
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        id: {type: integer}
                        error: {type: string}
                        code: {type: string, nullable: true}
        '400':
          description: Invalid ids or status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...
  /api/admin/tournaments/{tournament_id}/categories/{category_id}/standings:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
      - {name: category_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: Group tables of a category
      description: |
        ATP round-robin order: wins, matches played on court, head-to-head (two players tied), % of sets, % of games, then the organizer.
        A finished super tiebreak counts 1 set and 1 game to 0. A walkover is a win or a loss with no sets or games.
        A retirement keeps the sets and games played. A double walkover is a loss for both.
        Empty groups for knockout categories.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tables
          content:
            application/json:
              schema:
                type: object
                properties:
                  category: {$ref: '#/components/schemas/TournamentCategory'}
                  qualifiers_per_group: {type: integer}
                  groups:
                    type: array
                    items: {$ref: '#/components/schemas/TournamentGroupStandings'}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...
  /api/tournaments/{slug}:
    parameters:
      - {name: slug, in: path, required: true, schema: {type: string}}
    get:
      tags: [Tournaments]
      summary: A tournament with its categories (stage, progress and capacity of each)
      responses:
        '200':
          description: Tournament
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/PublicTournament'}
                  categories:
                    type: array
                    items: {$ref: '#/components/schemas/PublicCategorySummary'}
                  progress: {type: object, properties: {done: {type: integer}, total: {type: integer}}}
        '404':
          description: Unknown, draft or cancelled tournament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...